+++ components/forms/DatePicker.jsx
@@ -1,3 +1,4 @@
+import * as React from "react";
 import { Icon } from "../primitives/Icon.jsx";
 import { Popover } from "../overlays/Popover.jsx";
 import { Calendar } from "./Calendar.jsx";